refactor(schema): drop phantom manifest types; stop bricking gallery on corrupt source subfield
审计第三轮的契约层发现，减法 + 一处韧性修复：

- histogram/ImageMetadata 类型活在契约包里却从不跨越边界：photos-manifest.json
  从不含直方图，web 端的解压器是生产死代码。删掉 apps/web 的 histogram.ts
  及其测试；把 HistogramData 移到 builder 的 image/histogram.ts、ImageMetadata
  移到 image/processor.ts（各自唯一真实使用者），schema 从此只描述 manifest
  真正能装的东西。
- lenient 解析此前会因 manifest.source 子字段损坏而整体硬失败——而 source
  在全仓没有任何读者。validateEnvelope 加 strictSource 开关：严格路径
  （新产出门）不变，lenient 只要求 source 是对象或缺失，剩下的交给既有的
  sourceField.normalize 抢救成 { provider: "unknown" }。一条坏 source 不再
  能瘫痪画廊或构建缓存。
